+++ waves/abaqus/odb_extract.py
@@ -52,7 +52,7 @@ def get_parser():
     parser.add_argument('-r', '--odb-report-args',
                         dest='odb_report_args',
                         type=str,
-                        help='arguments to give to the odbreport command',
+                        help='Arguments to give to the odbreport command. Require the ``option=value`` interface style.',
                         metavar='"step=step1 results"')
     parser.add_argument('-a', '--abaqus-command',
                         dest='abaqus_command',
@@ -81,6 +81,11 @@ def odb_extract(input_file,
                 delete_report_file=False,
                 verbose=False):
     """The odb_extract Abaqus data extraction tool. Most users should use the associated command line interface.
+
+    .. warning::
+
+       ``odb_extract`` *requires* Abaqus arguments for ``odb_report_args`` in the form of ``option=value``, e.g.
+       ``step=step_name``.
 
     :param list input_file: A list of ``*.odb`` files to extract. Current implementation only supports extraction on the
         first file in the list.
@@ -113,14 +118,14 @@ def odb_extract(input_file,
                        f'Changing output file extension. Output file name {output_file}')
         file_suffix = output_type
     odb_report_args = odb_report_args
-    job_name = file_base_name
+    job_name = path_output_file.with_suffix('.csv')
     time_stamp = datetime.now().strftime(_settings._default_timestamp_format)
     if not odb_report_args:
-        odb_report_args = f'job={job_name} odb={input_file} all mode=CSV blocked'
+        odb_report_args = f"job={job_name.with_suffix('')} odb={input_file} all mode=CSV blocked"
     else:
         if 'odb=' in odb_report_args or 'job=' in odb_report_args:
             print_warning(verbose, f'Argument to odbreport cannot include odb or job. Will use default odbreport arguments.')
-            odb_report_args = f'job={job_name} odb={input_file} all mode=CSV blocked'
+            odb_report_args = f"job={job_name.with_suffix('')} odb={input_file} all mode=CSV blocked"
     if path_output_file.exists():
         new_output_file = f"{str(path_output_file.with_suffix(''))}_{time_stamp}.{file_suffix}"
         print_warning(verbose, f'{output_file} already exists. Will use {new_output_file} instead.')
@@ -131,7 +136,7 @@ def odb_extract(input_file,
     if 'odb=' not in odb_report_args:
         odb_report_args = f'odb={input_file} {odb_report_args.strip()}'
     if 'job=' not in odb_report_args:
-        odb_report_args = f'job={job_name} {odb_report_args.strip()}'
+        odb_report_args = f"job={job_name.with_suffix('')} {odb_report_args.strip()}"
     if 'blocked' not in odb_report_args:
         odb_report_args = f'{odb_report_args.strip()} blocked'
     if 'invariants' in odb_report_args:
@@ -151,7 +156,6 @@ def odb_extract(input_file,
         job_name = path_input_file
         call_odbreport = False
     else:
-        job_name = f'{file_base_name}{_settings._default_odbreport_extension}'
         call_odbreport = True
     if Path(job_name).exists() and not odbreport_file:
         call_odbreport = False  # Don't call odbreport again if the report file already exists
@@ -238,3 +242,7 @@ def main():
                          args.delete_report_file,
                          args.verbose,
             ))  # pragma: no cover
+
+
+if __name__ == "__main__":
+    sys.exit(main())