+++ waves/tests/test_builders.py
@@ -210,13 +210,16 @@ abaqus_extract_emitter_input = {
                       ['dummy.h5', 'dummy_datasets.h5', 'dummy.csv', 'dummy.h5.stdout']),
     'one target': (['new_name.h5'],
                    [source_file],
-                   ['new_name.h5', 'new_name_datasets.h5', 'dummy.csv', 'new_name.h5.stdout']),
+                   ['new_name.h5', 'new_name_datasets.h5', 'new_name.csv', 'new_name.h5.stdout']),
     'bad extension': (['new_name.txt'],
                       [source_file],
                       ['dummy.h5', 'new_name.txt', 'dummy_datasets.h5', 'dummy.csv', 'dummy.h5.stdout']),
     'subdirectory': (['set1/dummy.h5'],
                     [source_file],
-                    ['set1/dummy.h5', 'set1/dummy_datasets.h5', 'set1/dummy.csv', 'set1/dummy.h5.stdout'])
+                    ['set1/dummy.h5', 'set1/dummy_datasets.h5', 'set1/dummy.csv', 'set1/dummy.h5.stdout']),
+    'subdirectory new name': (['set1/new_name.h5'],
+                    [source_file],
+                    ['set1/new_name.h5', 'set1/new_name_datasets.h5', 'set1/new_name.csv', 'set1/new_name.h5.stdout'])
 }